move contractsHelper to helpers folder
