+++ src/helpers/contractsHelper.js
@@ -22,12 +22,12 @@ async function connectLatest(web3, abiFile) {
 }
 
 function getDeploysFile(networkId, contractName) {
-    const deploysFileName = `./abiniser/deployments/${networkId}/${contractName}_DEPLOYS.json`;
+    const deploysFileName = `src/abiniser/deployments/${networkId}/${contractName}_DEPLOYS.json`;
     let deploysFile;
 
     try {
         /* must provide fileName string again for webpack (needs to be statically analysable) */
-        deploysFile = require(`./abiniser/deployments/${networkId}/${contractName}_DEPLOYS.json`);
+        deploysFile = require(`src/abiniser/deployments/${networkId}/${contractName}_DEPLOYS.json`);
     } catch (error) {
         throw new Error(`Couldn't import deployment file ${deploysFileName} for ${contractName}\n${error}`);
     }
